test(fft): split CUDA source-contract checks
